refactor: add MemoCardContent view
