+++ MoeMemos.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -24,7 +24,7 @@
       "location" : "https://github.com/mudkipme/swift-markdown.git",
       "state" : {
         "branch" : "main",
-        "revision" : "ad0b81fed55c4a72fdaeb5535884bd2886de96f4"
+        "revision" : "2e0ffeea72a69c36eed2966528df30e75f3c2cf3"
       }
     },
     {
