+++ MoeMemos/Model/Memo.swift
@@ -60,3 +60,18 @@ extension MemosVisibility {
         }
     }
 }
+
+extension Memo {
+    func renderTime() -> String {
+        if Calendar.current.dateComponents([.day], from: createdTs, to: .now).day! > 7 {
+            let formatter = DateFormatter()
+            formatter.dateStyle = .long
+            formatter.timeStyle = .short
+            return formatter.string(from: createdTs)
+        }
+        
+        let formatter = RelativeDateTimeFormatter()
+        formatter.dateTimeStyle = .named
+        return formatter.localizedString(for: createdTs, relativeTo: .now)
+    }
+}
